broke down into components, weird auction error around router's subgraphs
