+++ packages/integration/ops/config/router/config.json
@@ -14,7 +14,7 @@
       "transactionManagerAddress": "0x8CdaF0CD259887258Bc13a92C0a6dA92698644C0"
     }
   },
-  "logLevel": "info",
+  "logLevel": "debug",
   "natsUrl": "nats://nats:4222",
   "authUrl": "http://auth:5040",
   "web3SignerUrl": "http://signer:9000",
